+++ storage/disk_examples.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+. disk_helpers.sh
+
+get_smallest_disk